iac/pyproject.toml
[tool.poetry]
name = "analytics-agent"
version = "0.0.1"
description = "AWS CDK infrastructure project for an analytics agent"
authors = ["Agent Team <agentteam@ai_alliance.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
aws-cdk-lib = "2.150.0"
constructs = ">=10.0.0,<11.0.0"

[tool.poetry.group.dev.dependencies]
coverage = "^7.2.2"
pytest = "^7.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
